### 2020-07-18 #2486 @whedon
[PRE REVIEW]: sbi - a toolkit for simulation-based inference
**Submitting author:** @janfb (<a href="http://orcid.org/0000-0003-4979-7092">Jan Boelts</a>)
**Repository:** <a href="https://github.com/mackelab/sbi" target ="_blank">https://github.com/mackelab/sbi</a>
**Version:** v0.10.0
