### 2020-07-22 #2505 @whedon
[REVIEW]: sbi - a toolkit for simulation-based inference
**Submitting author:** @janfb (<a href="http://orcid.org/0000-0003-4979-7092">Jan Boelts</a>)
**Repository:** <a href="https://github.com/mackelab/sbi" target ="_blank">https://github.com/mackelab/sbi</a>
**Version:** v0.10.0
