### 2020-07-23 f01205e
loop over rounds now part of API
