### 2020-07-23 961a338
loop over rounds now part of API
