+++ sbi/inference/snpe/snpe_base.py
@@ -5,7 +5,19 @@
 from abc import ABC, abstractmethod
 from copy import deepcopy
 from sbi.user_input.user_input_checks import check_estimator_arg
-from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
+from typing import (
+    Callable,
+    Optional,
+    Union,
+    Dict,
+    Any,
+    Tuple,
+    Union,
+    cast,
+    List,
+    Sequence,
+    TypeVar,
+)
 from warnings import warn
 
 import numpy as np
@@ -157,6 +169,8 @@ class PosteriorEstimator(NeuralInference, ABC):
             Posterior $p(\theta|x)$ that can be sampled and evaluated.
         """
 
+        # TODO: check the proposal. E.g. if it's a neuralPosterior, see if it has a default x.
+
         self._warn_if_retrain_from_scratch_snpe(retrain_from_scratch_each_round)
 
         # Calibration kernels proposed in Lueckmann, Gonçalves et al., 2017.
@@ -165,82 +179,78 @@ class PosteriorEstimator(NeuralInference, ABC):
 
         max_num_epochs = 2 ** 31 - 1 if max_num_epochs is None else max_num_epochs
 
-        num_sims_per_round = self._ensure_list(num_simulations_per_round, num_rounds)
+        self._maybe_update_round(proposal)
 
-        for round_, num_sims in enumerate(num_sims_per_round):
+        # Run simulations for the round.
+        theta, x = self._run_simulations(proposal, num_simulations)
+        self._append_to_data_bank(theta, x, self._round)
 
-            # Run simulations for the round.
-            theta, x = self._run_simulations(round_, num_sims)
-            self._append_to_data_bank(theta, x, round_)
+        # Load data from most recent round.
+        theta, x, _ = self._get_from_data_bank(self._round, exclude_invalid_x, False)
 
-            # Load data from most recent round.
-            theta, x, _ = self._get_from_data_bank(round_, exclude_invalid_x, False)
-
-            # First round or if retraining from scratch:
-            # Call the `self._build_neural_net` with the rounds' thetas and xs as
-            # arguments, which will build the neural network.
-            # This is passed into NeuralPosterior, to create a neural posterior which
-            # can `sample()` and `log_prob()`. The network is accessible via `.net`.
-            if round_ == 0 or retrain_from_scratch_each_round:
-                x_shape = x_shape_from_simulation(x)
-                self._posterior = NeuralPosterior(
-                    method_family="snpe",
-                    neural_net=self._build_neural_net(theta, x),
-                    prior=self._prior,
-                    x_shape=x_shape,
-                    sample_with_mcmc=self._sample_with_mcmc,
-                    mcmc_method=self._mcmc_method,
-                    mcmc_parameters=self._mcmc_parameters,
-                    get_potential_function=PotentialFunctionProvider(),
-                )
-                self._handle_x_o_wrt_amortization(x_o, x_shape, num_rounds)
-
-            # Fit posterior using newly aggregated data set.
-            self._train(
-                round_=round_,
-                training_batch_size=training_batch_size,
-                learning_rate=learning_rate,
-                validation_fraction=validation_fraction,
-                stop_after_epochs=stop_after_epochs,
-                max_num_epochs=cast(int, max_num_epochs),
-                clip_max_norm=clip_max_norm,
-                calibration_kernel=calibration_kernel,
-                exclude_invalid_x=exclude_invalid_x,
-                discard_prior_samples=discard_prior_samples,
+        # First round or if retraining from scratch:
+        # Call the `self._build_neural_net` with the rounds' thetas and xs as
+        # arguments, which will build the neural network.
[…]
