### 2020-07-24 74d88ce
check proposal distribution
