### 2020-07-24 f0b22f9
new multi-round interface adapted for snle
