### 2020-07-24 82e3352
new multi-round interface for snre
