### 2020-07-24 cf5eab9
new multi-round interface for snre
