### 2020-07-24 8f7051c
remove num_rounds from call made by simple interface
