### 2020-07-24 25d5df6
updated tests for new api
