### 2020-07-24 17e3cde
updated tests for new api
