+++ tests/linearGaussian_mdn.py
@@ -48,9 +48,7 @@ def mdn_inference_with_different_methods(method):
 
     infer = method(simulator, prior, density_estimator="mdn")
 
-    posterior = infer(
-        num_rounds=1, num_simulations_per_round=1000, training_batch_size=50
-    ).set_default_x(x_o)
+    posterior = infer(num_simulations=1000, training_batch_size=50).set_default_x(x_o)
     samples = posterior.sample((num_samples,))
 
     # Compute the c2st and assert it is near chance level of 0.5.
