### 2020-07-24 bc7fe1f
warn if multi-round snpe-c and invalid simulations
