+++ sbi/utils/__init__.py
@@ -1,19 +1,21 @@
 # flake8: noqa
-from sbi.utils.get_nn_models import classifier_nn, likelihood_nn
-from sbi.utils.get_nn_models import posterior_nn
+from sbi.user_input.user_input_checks import check_estimator_arg
+from sbi.user_input.user_input_checks_utils import MultipleIndependent
+from sbi.utils.get_nn_models import classifier_nn, likelihood_nn, posterior_nn
 from sbi.utils.io import get_data_root, get_log_root, get_project_root
 from sbi.utils.plot import pairplot
 from sbi.utils.sbiutils import (
-    standardizing_transform,
-    standardizing_net,
-    sample_posterior_within_prior,
-    del_entries,
     clamp_and_warn,
-    handle_invalid_x,
-    warn_on_invalid_x,
-    x_shape_from_simulation,
+    del_entries,
     get_data_since_round,
+    handle_invalid_x,
     mask_sims_from_prior,
+    sample_posterior_within_prior,
+    standardizing_net,
+    standardizing_transform,
+    warn_on_invalid_x,
+    warn_on_invalid_x_for_snpec_leakage,
+    x_shape_from_simulation,
 )
 from sbi.utils.torchutils import (
     BoxUniform,
@@ -41,6 +43,3 @@ from sbi.utils.typechecks import (
     is_positive_int,
     is_power_of_two,
 )
-
-from sbi.user_input.user_input_checks_utils import MultipleIndependent
-from sbi.user_input.user_input_checks import check_estimator_arg
