### 2020-07-24 4ba936a
if proposal is used in first round, we use atomic loss
