### 2020-07-24 2709f92
getting rid of x_o_trained_on
