### 2020-07-24 ab3f1b6
getting rid of x_o_trained_on
