### 2020-07-24 359461a
Adapting examples/tutorials
