### 2020-07-24 241cf69
added changes to CHANGELOG.md for v0.11.0
