+++ sbi/inference/base.py
@@ -1,8 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-from __future__ import annotations
-
 from abc import ABC
 from copy import deepcopy
 from datetime import datetime
