+++ sbi/utils/get_nn_models.py
@@ -1,7 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-from __future__ import annotations
 
 from typing import Any, Callable
 from sbi.neural_nets.flow import build_made, build_maf, build_nsf
