+++ sbi/user_input/user_input_checks.py
@@ -1,7 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-from __future__ import annotations
 
 import warnings
 from typing import Callable, Optional, Sequence, Tuple, Union, cast
