+++ sbi/inference/snle/snle_base.py
@@ -1,7 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-from __future__ import annotations
 
 from abc import ABC
 from typing import Callable, Dict, Optional, Union
