+++ sbi/inference/snre/snre_base.py
@@ -1,4 +1,3 @@
-from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from typing import Callable, Dict, Optional, Union
