+++ sbi/inference/abc/abc_base.py
@@ -1,4 +1,3 @@
-from __future__ import annotations
 
 from abc import ABC
 from typing import Callable, Union
