+++ sbi/inference/abc/smcabc.py
@@ -1,4 +1,3 @@
-from __future__ import annotations
 
 import logging
 from typing import Callable, Optional, Tuple, Union
