### 2020-07-26 166cc2f
Remove recursive type check
- Seems to be required for python < 3.7
