+++ sbi/inference/snle/snle_base.py
@@ -3,7 +3,7 @@
 
 
 from abc import ABC
-from typing import Callable, Dict, Optional, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast
 
 import numpy as np
 import torch
