+++ sbi/inference/snle/snle_a.py
@@ -2,7 +2,7 @@
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
 
-from typing import Callable, Optional, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast
 
 import torch
 from torch import Tensor
