+++ sbi/inference/snre/snre_a.py
@@ -1,5 +1,5 @@
 
-from typing import Callable, Optional, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union
 
 import torch
 from torch import Tensor, nn, ones
