### 2020-07-26 631b4a7
Fix typing imports
- Import all used common types in same way for each file
