+++ sbi/utils/get_nn_models.py
@@ -2,7 +2,7 @@
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
 
-from typing import Any, Callable
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List
 from sbi.neural_nets.flow import build_made, build_maf, build_nsf
 from sbi.neural_nets.mdn import build_mdn
 from sbi.neural_nets.classifier import (
