+++ sbi/inference/snre/snre_base.py
@@ -1,6 +1,6 @@
 
 from abc import ABC, abstractmethod
-from typing import Callable, Dict, Optional, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List
 
 import numpy as np
 import torch
