+++ sbi/inference/base.py
@@ -5,7 +5,7 @@ from abc import ABC
 from copy import deepcopy
 from datetime import datetime
 from pathlib import Path
-from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List
 from warnings import warn
 
 import torch
