### 2020-07-26 bbf0b2e
Other missing types
