+++ sbi/inference/abc/mcabc.py
@@ -1,5 +1,5 @@
 
-from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 
 import torch
 from numpy import ndarray
