+++ sbi/user_input/user_input_checks_utils.py
@@ -3,7 +3,7 @@
 
 
 import warnings
-from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 
 import torch
 from scipy.stats._distn_infrastructure import rv_frozen
