### 2020-07-26 103f1f4
Install compatible nflows version in setup.py
