### 2020-07-26 7076536
Upgrade compatibility in README
