### 2020-07-26 afab341
Switch to pip version of nflows
