### 2020-07-27 e963c29
Add missing type imports + make compatible for Python 3.6 (#275)
