+++ sbi/utils/plot.py
@@ -3,11 +3,11 @@
 
 import collections
 import inspect
-from typing import Optional, Tuple, Union, List
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 
 import torch
 import matplotlib as mpl
-import matplotlib.pyplot as plt
+from matplotlib import pyplot as plt
 import numpy as np
 import six
 from scipy.stats import gaussian_kde
