+++ sbi/inference/abc/mcabc.py
@@ -1,6 +1,5 @@
-from __future__ import annotations
 
-from typing import Callable, Optional, Tuple, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 
 import torch
 from numpy import ndarray
