+++ sbi/user_input/user_input_checks_utils.py
@@ -1,10 +1,9 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-from __future__ import annotations
 
 import warnings
-from typing import Optional, Sequence, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 
 import torch
 from scipy.stats._distn_infrastructure import rv_frozen
@@ -171,7 +170,7 @@ class MultipleIndependent(Distribution):
 
     def __init__(
         self, dists: Sequence[Distribution], validate_args=None,
-    ) -> MultipleIndependent:
+    ):
         self._check_distributions(dists)
 
         self.dists = dists
