+++ sbi/utils/torchutils.py
@@ -7,10 +7,10 @@ import numpy as np
 import torch
 from torch import Tensor, float32, device
 from torch.distributions import Independent, Uniform
-from typing import Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 import warnings
 
-import sbi.utils as utils
+from sbi import utils as utils
 from sbi.types import Array, OneOrMore, ScalarFloat
 
 
