+++ sbi/inference/abc/abc_base.py
@@ -1,7 +1,6 @@
-from __future__ import annotations
 
 from abc import ABC
-from typing import Callable, Union
+from typing import Callable, Optional, Union, Dict, Any, Tuple, Union, cast, List, Sequence, TypeVar
 
 import torch
 from torch import Tensor
