+++ sbi/__version__.py
@@ -1,6 +1,6 @@
 # This file is part of sbi, a toolkit for simulation-based inference. sbi is licensed
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
-VERSION = (0, 10, 2)
+VERSION = (0, 11, 0)
 
 __version__ = '.'.join(map(str, VERSION))
