### 2020-07-30 9be1801
fixed log_prob bug in snre
