### 2020-07-30 b6d7682
subclassing posterior
