+++ sbi/inference/snle/snle_a.py
@@ -2,24 +2,11 @@
 # under the Affero General Public License v3, see <https://www.gnu.org/licenses/>.
 
 
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    List,
-    Optional,
-    Sequence,
-    Tuple,
-    TypeVar,
-    Union,
-    cast,
-)
+from typing import Any, Callable, Dict, Optional, Union
 
-import torch
-from torch import Tensor
 from torch.utils.tensorboard import SummaryWriter
 
-from sbi.inference.posterior import NeuralPosterior
+from sbi.inference.posteriors.posterior import NeuralPosterior
 from sbi.inference.snle.snle_base import LikelihoodEstimator
 from sbi.utils import del_entries
 
