+++ sbi/inference/snpe/snpe_base.py
@@ -4,18 +4,7 @@
 
 from abc import ABC, abstractmethod
 from copy import deepcopy
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    List,
-    Optional,
-    Sequence,
-    Tuple,
-    TypeVar,
-    Union,
-    cast,
-)
+from typing import Any, Callable, Dict, Optional, Union, cast
 from warnings import warn
 
 import numpy as np
@@ -28,9 +17,8 @@ from torch.utils.tensorboard import SummaryWriter
 
 from sbi import utils as utils
 from sbi.inference import NeuralInference
-from sbi.inference.posterior import NeuralPosterior
+from sbi.inference.posteriors.snpe_posterior import SnpePosterior
 from sbi.types import ScalarFloat
-from sbi.user_input.user_input_checks import check_estimator_arg
 from sbi.utils import check_estimator_arg, x_shape_from_simulation
 
 
@@ -125,7 +113,7 @@ class PosteriorEstimator(NeuralInference, ABC):
         exclude_invalid_x: bool = True,
         discard_prior_samples: bool = False,
         retrain_from_scratch_each_round: bool = False,
-    ) -> NeuralPosterior:
+    ) -> SnpePosterior:
         r"""Run SNPE.
 
         Return posterior $p(\theta|x)$ after inference.
@@ -185,7 +173,7 @@ class PosteriorEstimator(NeuralInference, ABC):
         # can `sample()` and `log_prob()`. The network is accessible via `.net`.
         if self._posterior is None or retrain_from_scratch_each_round:
             x_shape = x_shape_from_simulation(x)
-            self._posterior = NeuralPosterior(
+            self._posterior = SnpePosterior(
                 method_family="snpe",
                 neural_net=self._build_neural_net(theta, x),
                 prior=self._prior,
