### 2020-07-30 e489775
subclassing posterior
