### 2020-07-30 f21d910
Fixes log_prob bug in SNRE (#280)
