### 2020-07-31 cfc1487
unified log_prob() interface
