### 2020-07-31 10c9bc2
unified log_prob() interface
