### 2020-07-31 abe2c15
Adapted docstrings to new posterior subclassing
