### 2020-07-31 b22706f
Adapted docstrings to new posterior subclassing
