### 2020-07-31 b40a401
inheritance in docstrings
