### 2020-07-31 2e80302
inheritance in docstrings
