### 2020-08-03 7b930eb
posterior base is now abstract base class
