### 2020-08-03 1321a89
updated posterior description
