### 2020-08-03 c67abce
updated posterior description
