+++ sbi/inference/posteriors/likelihood_based_posterior.py
@@ -24,17 +24,13 @@ from sbi.utils import del_entries
 from sbi.utils.torchutils import atleast_2d_float32_tensor
 
 
-class SNRE_Posterior(NeuralPosterior):
+class LikelihoodBasedPosterior(NeuralPosterior):
     r"""Posterior $p(\theta|x)$ with `log_prob()` and `sample()` methods, obtained with
-    SNRE.<br/><br/>
-    SNRE trains a neural network to approximate likelihood ratios, which in turn can be
-    used obtain an unnormalized posterior $p(\theta|x) \propto p(x|\theta) \cdot
-    p(\theta)$. The `SNRE_Posterior` class wraps the trained network such that one can
-    directly evaluate the unnormalized posterior log-probability $p(\theta|x) \propto
-    p(x|\theta) \cdot p(\theta)$ and draw samples from the posterior with
-    MCMC. Note that, in the case of single-round SNRE_A / AALR, it is possible to
-    evaluate the log-probability of the **normalized** posterior, but sampling still
-    requires MCMC.<br/><br/>
+    SNLE.<br/><br/>
+    SNLE trains a neural network to approximate the likelihood $p(x|\theta)$. The
+    `SNLE_Posterior` class wraps the trained network such that one can directly evaluate
+    the unnormalized posterior log probability $p(\theta|x) \propto p(x|\theta) \cdot
+    p(\theta)$ and draw samples from the posterior with MCMC.<br/><br/>
     The neural network itself can be accessed via the `.net` attribute.
     """
 
@@ -78,9 +74,7 @@ class SNRE_Posterior(NeuralPosterior):
         r"""
         Returns the log-probability of $p(x|\theta) \cdot p(\theta).$
 
-        This corresponds to an **unnormalized** posterior log-probability. Only for
-        single-round SNRE_A / AALR, the returned log-probability will correspond to the
-        **normalized** log-probability.
+        This corresponds to an **unnormalized** posterior log-probability.
 
         Args:
             theta: Parameters $\theta$.
@@ -101,24 +95,12 @@ class SNRE_Posterior(NeuralPosterior):
 
         theta, x = self._prepare_theta_and_x_for_log_prob_(theta, x)
 
-        self._warn_log_prob_snre()
+        warn(
+            "The log probability from SNL is only correct up to a normalizing constant."
+        )
 
         with torch.set_grad_enabled(track_gradients):
-            log_ratio = self.net(torch.cat((theta, x), dim=1)).reshape(-1)
-            return log_ratio + self._prior.log_prob(theta)
-
-    def _warn_log_prob_snre(self) -> None:
-        if self._method_family == "snre_a":
-            if self._num_trained_rounds > 1:
-                warn(
-                    "The log-probability from AALR / SNRE-A beyond round 1 is only"
-                    " correct up to a normalizing constant."
-                )
-        elif self._method_family == "snre_b":
-            warn(
-                "The log probability from SNRE_B is only correct up to a normalizing "
-                "constant."
-            )
+            return self.net.log_prob(x, theta) + self._prior.log_prob(theta)
 
     def sample(
         self,
