+++ sbi/inference/posteriors/base_posterior.py
@@ -513,7 +513,7 @@ class NeuralPosterior(ABC):
         return x_matched
 
     def __repr__(self):
-        desc = f"""NeuralPosterior(
+        desc = f"""{self.__class__.__name__}(
                method_family={self._method_family},
                net=<a {self.net.__class__.__name__}, see `.net` for details>,
                prior={self._prior!r},
@@ -527,17 +527,39 @@ class NeuralPosterior(ABC):
         focused_msg = "multi-round"
 
         default_x_msg = (
-            f" Evaluates and samples by default at x={self.default_x.tolist()!r}"
+            f" Evaluates and samples by default at x={self.default_x.tolist()!r}."
             if self.default_x is not None
             else ""
         )
 
+        purpose_leakage = (
+            "It allows to .sample() and .log_prob() the posterior"
+            " and wraps the output of the .net to avoid leakage into regions with 0"
+            " prior probability."
+        )
+
+        normalized_or_not = (
+            ""
+            if (self._method_family == "snre_a" and self._num_trained_rounds == 1)
+            else "_unnormalized_ "
+        )
+        purpose_mcmc = (
+            f"It provides MCMC to .sample() from the posterior and "
+            f"can evaluate the {normalized_or_not}posterior density with .log_prob()."
+        )
+
+        purpose = purpose_leakage if self._method_family == "snpe" else purpose_mcmc
+
+        # The net might be sequential because it can have a standardization net. Hence,
+        # we only access its last entry if it is a nn.Sequential.
+        actual_net = self.net[-1] if isinstance(self.net, nn.Sequential) else self.net
         desc = (
             f"Posterior conditional density p(θ|x) "
-            f"({msg.get(self._num_trained_rounds, focused_msg)}).{default_x_msg}.\n\n"
-            f"This neural posterior was obtained with a "
+            f"({msg.get(self._num_trained_rounds, focused_msg)}).{default_x_msg}\n\n"
+            f"This {self.__class__.__name__}-object was obtained with a "
             f"{self._method_family.upper()}-class "
-            f"method using a {self.net.__class__.__name__.lower()}."
+            f"method using a {actual_net.__class__.__name__.lower()}.\n"
+            f"{purpose}"
         )
 
         return desc
