### 2020-08-03 d416410
purpose moved to child classes
