### 2020-08-03 5900b3d
purpose moved to child classes
