### 2020-08-03 8c3389c
prepare for sample method
