+++ sbi/inference/posteriors/likelihood_based_posterior.py
@@ -143,14 +143,8 @@ class LikelihoodBasedPosterior(NeuralPosterior):
             Samples from posterior.
         """
 
-        x = atleast_2d_float32_tensor(self._x_else_default_x(x))
-        self._ensure_single_x(x)
-        self._ensure_x_consistent_with_default_x(x)
-        num_samples = torch.Size(sample_shape).numel()
-
-        mcmc_method = mcmc_method if mcmc_method is not None else self.mcmc_method
-        mcmc_parameters = (
-            mcmc_parameters if mcmc_parameters is not None else self.mcmc_parameters
+        x, num_samples, mcmc_method, mcmc_parameters = self._prepare_for_sample(
+            x, sample_shape, mcmc_method, mcmc_parameters
         )
 
         samples = self._sample_posterior_mcmc(
