### 2020-08-04 d556ebd
get actual_net more robustly
