### 2020-08-04 a2cf749
remove superfluous check for snpe in sample_with_mcmc
