### 2020-08-04 715a121
ensure correct round number if presimulated was provided
