### 2020-08-04 32d08dd
Update versions in environment.yml
