+++ environment.yml
@@ -27,9 +27,9 @@ dependencies:
   - pip
   - pip:
     - "pyknos==0.11"
-    - pyro-ppl
+    - "pyro-ppl==1.3.1"
     - -e ".[dev]"
-  - "python >= 3.7.0"
+  - "python >= 3.6.0"
   - "pytorch >= 1.5.1"
   - scikit-learn
   - scipy
