### 2020-08-07 dfd2af6
Detach weights for SIR immediately to save memory
