### 2020-08-07 b70899a
Move StandardizeInputs outside of function for pickling (#291)
