+++ docs/README.md
@@ -14,3 +14,10 @@ The docs can be updated to GitHub using:
 ```bash
 jupyter nbconvert --to markdown ../tutorial/*.ipynb --output-dir docs/tutorial/ && mkdocs gh-deploy
 ```
+
+## Contributing FAQ
+
+Create a new markdown file named `question_XX.md` in the `docs/faq` folder, where `XX` 
+is a running index for the questions. The file should start with the question as title 
+(i.e. starting with a `#`) and then have the answer below. Additionally, you need to 
+add a link to your question in the markdown file `docs/faq.md`.
