+++ docs/docs/faq.md
@@ -1,0 +1,5 @@
+# Frequently asked questions
+
+[Can the algorithms deal with invalid data, e.g. NaN or inf?](faq/question_02.md)
+
+[What should I do when my 'posterior samples are outside of the prior support' in SNPE?](faq/question_01.md)
