### 2020-08-10 2e715c4
update required python version on website
