### 2020-08-12 #2505 @EiffL
Everything looks great, install is a breeze, runs on colab, nice documentation, nice test suite, very useful package!
Very happy to accept, and many thanks to the team for this tool :-)
