### 2020-08-12 #2505 @cranmer
Hello,

Sorry for the delay, a few unexpected things fell on me.

I have time now, but my invitation has expired.

Kyle

> On Aug 8, 2020, at 9:07 AM, Dan Foreman-Mackey <notifications@github.com> wrote:
>
>
> @EiffL <https://github.com/EiffL>, @cranmer <https://github.com/cranmer>: I just wanted to check in with you guys to keep this review on your radar. Let me know if you have any questions!
>
