### 2020-08-12 #2505 @cranmer
The check list has:

> **Contribution and authorship:** Has the submitting author (@janfb) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

I wasn't sure where to find what is considered the “paper authors” at first, and then I found the PDF for the article proof.

Misc feedback for JOSS would be to have link to the PDF in the checklist. (Or maybe it would normally be there and it just failed for some reason).
