### 2020-08-12 #2505 @cranmer
I know that Jakob H. Macke is part of this effort and belongs on author list (GitHub organization is Mackelab) but noting it since he doesn't show up on contributor list.
