### 2020-08-12 mackelab/sbi#298 @cranmer
Summarizing network / learned summary statistics
From the API documentation, it seems like the package is restricted to simulators that produce output `x` with a constant shape. I wondered about simulators that produce output of varying size and shape. Currently, it seems this would possible if one introduces a function that wraps the simulator to produce fixed length summary statistics. This was not immediately clear to me from the API documentation or the [home](https://www.mackelab.org/sbi/) page in the documentation.

 I do see that the Hodgkin-Huxley example does exactly this with domain-motivated summary statistics (as opposed to learned summary statistics).

However, the software paper mentions

> Moreover, if dimensionality reduction of the simulator output is desired, sbi can use a trainable summarizing network to extract relevant features from raw simulator output and spare the user manual feature engineering.

I can see how this wold work with a pre-trained network that produces summary statistics, but my reading of that sentence with the word "trainable" is that SBI can also train the summarizing network. I haven't found this yet in the API documentation. If it is possible, an example would be great. Either way it would be nice to clarify this in the documentation and/or reword the sentence in the software paper.
