### 2020-08-12 #2505 @cranmer
Ok, I'm basically done. I don't know if I needed to explicitly run all the tests, they take a while. I did install the environment locally and can confirm that the tests are running.

The package is very nice, the interface to simulators seems natural, and I was able to run tutorial examples! The examples themselves are well chosen.

I created a few issues recommending improvements in documentation, but I don't see them as show stoppers.

There is one issue that had me uncheck the **Functionality documentation:**:
https://github.com/mackelab/sbi/issues/298
The package would be useful even if this isn't possible, but the software paper says it is, and I haven't been able to confirm that yet. If it isn't possible, then I think a small tweak to the sentence in the software paper would be sufficient for publication.
