### 2020-08-12 mackelab/sbi#298 @dfm
Linking to https://github.com/openjournals/joss-reviews/issues/2505 for posterity
