### 2020-08-13 64e5ea6
pass arguments to nn build fn explicitly for api reference
