### 2020-08-13 2b53f98
added documentation to build nn functions
