### 2020-08-13 mackelab/sbi#298 @michaeldeistler
Good point, thanks for raising this!

It is indeed possible to use a trainable network by passing an `embedding_net` to e.g. `posterior_nn` [here](https://github.com/mackelab/sbi/blob/7c56cc210f226839bf91d037ccd258a733096de5/sbi/utils/get_nn_models.py#L48). The `embedding_net` can be any trainable `torch.nn.Module`. It is also possible to pass such a network for SNRE.

This argument got lost from the the documentation page since we started using `**kwargs` [here](https://github.com/mackelab/sbi/blob/7c56cc210f226839bf91d037ccd258a733096de5/sbi/utils/get_nn_models.py#L48) at some point. We will fix this in #299. The PR will also add a bit more explanation on the `embedding_net`.

Writing an example for the `embedding_net` is on our hot desk, and is also documented in #279.
