### 2020-08-13 bbca289
allow for embedding_net to map onto lower dimension
