### 2020-08-13 7858273
allow for embedding_net to map onto lower dimension
