+++ sbi/neural_nets/mdn.py
@@ -33,7 +33,8 @@ def build_mdn(
         Neural network.
     """
     x_numel = batch_x[0].numel()
-    y_numel = batch_y[0].numel()
+    # Infer the output dimensionality of the embedding_net by making a forward pass.
+    y_numel = embedding_net(batch_y[:1]).numel()
 
     transform = transforms.IdentityTransform()
 
