### 2020-08-13 95be13a
Non-atomic variant of APT / SNPE-C
