### 2020-08-13 5c23f43
Non-atomic variant of APT / SNPE-C
