### 2020-08-13 #2505 @cranmer
Thanks to @michaeldeistler of SBI team to quickly address problem. There is a PR that will fix the API documentation, and an example is planned to follow @dfm I'm not really sure about the procedure here. Do we wait for the PR to be merged and new API docs to be generated, or do we proceed assuming that it's in progress and on the way? (The example demonstrating this feature is nice, but I don't think it is essential).

As I wrote earlier, I think SBI is great and the submission is publishable even without this feature being there, except for that it is explicitly called out in the software paper so it should be in the API docs.
