### 2020-08-13 #2505 @dfm
@cranmer: Thanks for the update! If you think that the functionality claims in the paper are sufficiently well justified once the `embedding_net`, etc. parameters are documented then I think it would be fine to proceed once that pull request has been merged without waiting for the full example in the docs.
