### 2020-08-13 ec7dfce
fix bug with additional kwargs
