+++ sbi/utils/get_nn_models.py
@@ -49,7 +49,7 @@ def classifier_nn(
             the network.
         hidden_features: Number of hidden features.
         embedding_net_theta:  Optional embedding network for parameters $\theta$.
-        embedding_net_x:  Optional embedding network for simulation outputs $x. This
+        embedding_net_x:  Optional embedding network for simulation outputs $x$. This
             embedding net allows to learn features from potentially high-dimensional
             simulation outputs.
     """
@@ -112,7 +112,9 @@ def likelihood_nn(
         z_score_x: Whether to z-score simulation outputs $x$ before passing them into
             the network.
         hidden_features: Number of hidden features.
-        num_transforms: Number of transforms when a flow is used.
+        num_transforms: Number of transforms when a flow is used. Only relevant if
+            density estimator is a normalizing flow (i.e. currently either a `maf` or a
+            `nsf`). Ignored if density estimator is a `mdn` or `made`.
         embedding_net: Optional embedding network for parameters $\theta$.
     """
 
@@ -125,7 +127,7 @@ def likelihood_nn(
                 "num_transforms",
                 "embedding_net",
             ),
-            (z_score_x, z_score_theta, hidden_features, num_transforms, embedding_net),
+            (z_score_x, z_score_theta, hidden_features, num_transforms, embedding_net,),
         )
     )
 
@@ -166,7 +168,9 @@ def posterior_nn(
         z_score_x: Whether to z-score simulation outputs $x$ before passing them into
             the network.
         hidden_features: Number of hidden features.
-        num_transforms: Number of transforms when a flow is used.
+        num_transforms: Number of transforms when a flow is used. Only relevant if
+            density estimator is a normalizing flow (i.e. currently either a `maf` or a
+            `nsf`). Ignored if density estimator is a `mdn` or `made`.
         embedding_net: Optional embedding network for simulation outputs $x$. This
             embedding net allows to learn features from potentially high-dimensional
             simulation outputs.
@@ -181,7 +185,7 @@ def posterior_nn(
                 "num_transforms",
                 "embedding_net",
             ),
-            (z_score_theta, z_score_x, hidden_features, num_transforms, embedding_net),
+            (z_score_theta, z_score_x, hidden_features, num_transforms, embedding_net,),
         )
     )
 
