### 2020-08-17 9938e3d
re-order installation instructions.
