### 2020-08-17 mackelab/sbi#298 @cranmer
👋 I see that #299 is merged. I think the API documentation online is still the same. Once that API documentation is updated and exposes `embedding_net` I will approve the JOSS submission. The example can come later and is not needed for JOSS paper.
