### 2020-08-18 49eb755
updated faq for avoiding leakage -> non-atomic APT
