### 2020-08-18 1ce771b
updated faq for avoiding leakage -> non-atomic APT
