+++ docs/docs/faq/question_01.md
@@ -15,5 +15,12 @@ This will make sampling slower, but samples will not 'leak'.
 
 - resort to single-round SNPE and (if necessary) increase your simulation budget.  
 
+- if your prior is either Gaussian (torch.distributions.multivariateNormal) or Uniform 
+(sbi.utils.BoxUniform), you can avoid leakage by using a mixture density network as 
+density estimator. I.e., using the 
+[flexible interface](https://www.mackelab.org/sbi/tutorial/03_flexible_interface/), set 
+`density_estimator='mdn'`. When running inference, there should be a print statement 
+"Using SNPE-C with non-atomic loss"
+
 - use a different algorithm, e.g. SNRE and SNLE. Note, however, that these algorithms
 can have different issues and potential pitfalls.  
