### 2020-08-18 mackelab/sbi#298 @cranmer
@michaeldeistler just want to make sure you see note above.
