### 2020-08-18 mackelab/sbi#298 @michaeldeistler
Hi Kyle, thanks, saw it. We'll make a new PyPI release soon (hopefully tomorrow) and then update the docs :)
