### 2020-08-19 5f7c59b
DOC: fix the url to the running tests in readme
