### 2020-08-19 ad81eb3
cast 1d uniform priors to boxuniform.
