### 2020-08-19 88d170a
cast 1d uniform priors to boxuniform.
