### 2020-08-19 ce4a6f2
new warning leakage in log_prob()
