### 2020-08-19 #2505 @janfb
@cranmer @dfm Please note we just released a new version of `sbi` adressing the above issues. The `docs` should be up to date now as well.
