### 2020-08-19 d1be55c
fix num_blocks bug
