+++ paper.bib
@@ -1,11 +1,21 @@
-@misc{abcpy-repo,
-  author       = {ABCpy authors, The},
-  title        = {ABCpy},
-  year         = {2017},
-  publisher    = {GitHub},
-  journal      = {GitHub repository},
-  howpublished = {\url{https://github.com/eth-cscs/abcpy}}
-}
+@inproceedings{dutta2017,
+ author = {Dutta, Ritabrata and Schoengens, Marcel and Onnela, Jukka-Pekka and Mira, Antonietta},
+ title = {{ABCpy}: {A} User-Friendly, Extensible, and Parallel Library for Approximate Bayesian Computation},
+ booktitle = {Proceedings of the Platform for Advanced Scientific Computing Conference},
+ series = {PASC '17},
+ year = {2017},
+ isbn = {978-1-4503-5062-4},
+ location = {Lugano, Switzerland},
+ pages = {8:1--8:9},
+ articleno = {8},
+ numpages = {9},
+ url = {http://doi.acm.org/10.1145/3093172.3093233},
+ doi = {10.1145/3093172.3093233},
+ acmid = {3093233},
+ publisher = {ACM},
+ address = {New York, NY, USA},
+ keywords = {ABC, Library, Parallel, Spark},
+} 
 
 @inproceedings{baydin2020,
   title     = {Etalumis: bringing probabilistic programming to scientific simulators at scale},
@@ -17,7 +27,7 @@
 }
 
 @article{beaumont2009,
-  title     = {Adaptive approximate Bayesian computation},
+  title     = {Adaptive approximate {B}ayesian computation},
   author    = {Beaumont, Mark A and Cornuet, Jean-Marie and Marin, Jean-Michel and Robert, Christian P},
   journal   = {Biometrika},
   volume    = {96},
@@ -47,7 +57,7 @@
 
 @misc{delfi-repo,
   author       = {mackelab.org},
-  title        = {DELFI: Density Estimation Likelihood-Free Inference},
+  title        = {{DELFI}: Density Estimation Likelihood-Free Inference},
   year         = {2017},
   publisher    = {GitHub},
   journal      = {GitHub repository},
@@ -73,7 +83,7 @@
 }
 
 @article{elfi2018,
-  title     = {ELFI: engine for likelihood-free inference},
+  title     = {{ELFI}: Engine for likelihood-free inference},
   author    = {Lintusaari, Jarno and Vuollekoski, Henri and Kangasr{\"a}{\"a}si{\"o}, Antti and Skyt{\'e}n, Kusti and J{\"a}rvenp{\"a}{\"a}, Marko and Marttinen, Pekka and Gutmann, Michael U and Vehtari, Aki and Corander, Jukka and Kaski, Samuel},
   journal   = {The Journal of Machine Learning Research},
   volume    = {19},
@@ -129,7 +139,7 @@
 
 @inproceedings{hermans2019,
   author    = {Hermans, Joeri and Begy, Volodimir and Louppe, Gilles},
-  title     = {Likelihood-free MCMC with Approximate Likelihood Ratios},
+  title     = {Likelihood-free {MCMC} with Approximate Likelihood Ratios},
   year      = {2020},
   booktitle = {Proceedings of the 37th International Conference on Machine Learning},
   volume    = {98},
@@ -157,7 +167,7 @@
 
 
 @article{Klinger2018,
-  title     = {pyABC: distributed, likelihood-free inference},
+  title     = {{pyABC}: distributed, likelihood-free inference},
   author    = {Klinger, Emmanuel and Rickert, Dennis and Hasenauer, Jan},
   journal   = {Bioinformatics},
   volume    = {34},
@@ -196,7 +206,7 @@
   number    = {1},
   pages     = {11},
[…]
