### 2020-08-20 7b3d13d
Update paper.md
