+++ paper.bib
@@ -2,19 +2,12 @@
  author = {Dutta, Ritabrata and Schoengens, Marcel and Onnela, Jukka-Pekka and Mira, Antonietta},
  title = {{ABCpy}: {A} User-Friendly, Extensible, and Parallel Library for Approximate Bayesian Computation},
  booktitle = {Proceedings of the Platform for Advanced Scientific Computing Conference},
- series = {PASC '17},
  year = {2017},
- isbn = {978-1-4503-5062-4},
- location = {Lugano, Switzerland},
  pages = {8:1--8:9},
  articleno = {8},
  numpages = {9},
  url = {http://doi.acm.org/10.1145/3093172.3093233},
- doi = {10.1145/3093172.3093233},
- acmid = {3093233},
- publisher = {ACM},
- address = {New York, NY, USA},
- keywords = {ABC, Library, Parallel, Spark},
+ doi = {10.1145/3093172.3093233}
 } 
 
 @inproceedings{baydin2020,
@@ -307,7 +300,7 @@
 }
 
 @article{pritchard1999,
-  title     = {Population growth of human Y chromosomes: a study of Y chromosome microsatellites.},
+  title     = {Population growth of human {Y} chromosomes: a study of {Y} chromosome microsatellites.},
   author    = {Pritchard, Jonathan K and Seielstad, Mark T and Perez-Lezaun, Anna and Feldman, Marcus W},
   journal   = {Molecular biology and evolution},
   volume    = {16},
@@ -337,7 +330,7 @@
 }
 
 @article{sisson2007,
-  title     = {Sequential monte carlo without likelihoods},
+  title     = {Sequential {Monte} {Carlo} without likelihoods},
   author    = {Sisson, Scott A and Fan, Yanan and Tanaka, Mark M},
   journal   = {Proceedings of the National Academy of Sciences},
   volume    = {104},
