### 2020-08-20 bdd7b2a
Update paper.md
