### 2020-08-20 #2505 @dfm
@janfb: Looks good! I just opened a pull request with some formatting edits for the manuscript. The following steps for you:

- [x] Accept that pull request
- [ ] Comment `@whedon generate pdf from branch joss-submission` on this issue thread
- [ ] Check the proof that `@whedon` generates to make sure everything looks good to you (it's harder to change later!)
- [ ] It looks like you just bumped the version number and made a release, so please confirm that `0.12.1` is the current version and make an archive of this release (on Zenodo or similar) where **the metadata (title & authors) exactly matches this paper** and report back with the DOI
