### 2020-08-20 #2505 @janfb
Regarding the version, yes, the most recent one is 0.12.1.

Two questions
1) we already have a version on arxiv: https://arxiv.org/abs/2007.09114 Is OK if we just make a new version of this submission with the most recent version of the paper?

2) The JOSS version of the article already has a DOI (10.21105/joss.02505) no? Or shall I create a new DOI using Zenodo?
