### 2020-08-20 #2505 @janfb
I double-checked the article -- all looks good!
I created a DOI using Zenodo and upload the zipped code of v0.12.1 (and not the paper PDF), the DOI is [10.5281/zenodo.3993098](https://doi.org/10.5281/zenodo.3993098) Is this how it should be?
